hotfix: FileSize is now long

+++ src/main/java/com/cadiducho/telegrambotapi/Animation.java
@@ -62,5 +62,5 @@ public class Animation {
     /**
      * Optional. File size
      */
-    @Json(name = "file_size") private Integer fileSize;
+    @Json(name = "file_size") private Long fileSize;
 }
