+++ src/main/java/com/cadiducho/telegrambotapi/Sticker.java
@@ -68,6 +68,6 @@ public class Sticker {
     /**
      * Optional. File size
      */
-    @Json(name = "file_size") private Integer fileSize;
+    @Json(name = "file_size") private Long fileSize;
 
 }
