+++ src/main/java/com/cadiducho/telegrambotapi/VideoNote.java
@@ -47,6 +47,6 @@ public class VideoNote {
     /**
      * Optional. File size
      */
-    @Json(name = "file_size") private Integer fileSize;
+    @Json(name = "file_size") private Long fileSize;
     
 }
